changed code to read in from text files

+++ src/main/java/org/petuum/lda/training/WordSortReducer.java
@@ -3,6 +3,7 @@ package org.petuum.lda.training;
 import java.io.IOException;
 import java.util.Iterator;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.JobConf;
@@ -11,7 +12,7 @@ import org.apache.hadoop.mapred.Reducer;
 import org.apache.hadoop.mapred.Reporter;
 
 public class WordSortReducer implements
-		Reducer<IntWritable, Text, Text, IntWritable> {
+		Reducer<DoubleWritable, Text, Text, DoubleWritable> {
 
 	@Override
 	public void configure(JobConf arg0) {
@@ -25,8 +26,8 @@ public class WordSortReducer implements
 	}
 
 	@Override
-	public void reduce(IntWritable arg0, Iterator<Text> arg1,
-			OutputCollector<Text, IntWritable> arg2, Reporter arg3)
+	public void reduce(DoubleWritable arg0, Iterator<Text> arg1,
+			OutputCollector<Text, DoubleWritable> arg2, Reporter arg3)
 			throws IOException {
 		while (arg1.hasNext()) {
 			arg2.collect(arg1.next(), arg0);
